Use the configured separator to combine slugs even _before_ postBuild() callback.
Reason: concatenating multiple-field slugs with spaces does not seem too
practical as then the various handler-callback have no chance to have a
look at the components -- it is not clear whether a space is really a
separator or just contained in one of the constituing fields.

Arguably, it would be even more practical to just pass the array of
components to Handler::onChangeDecision() and Handler::postSlugBuild()
but this would probably too big a change.

So if we have two field firstName = "James Tiberius" and
lastName="Kirk", then without this change the generated slug passed to
the handler's callbacks is "James Tiberius Kirk". With this change it
will be "James Tiberius-Kirk", given the default separator "-" is used.

This allows the callbacks to have a look at the components.