SoftDeletable: optionally cascade soft-deletion and undeletion.
Strategy is to mark association properties as cascadable with a property
annotation. Then propagate the soft-deletion down to the associated
entities.

On undelelete (setting the delete-stamp to null) any associated cascaded
entity which suffered soft-deletetion at a later point in time is
undeleted.

Additionally, there are soft-undelete events.

It has to be noted that cascading a soft-deletion should not be necessary
as this can already be achieved by orphanRemoval and cascade=delete options,
if the orphan or associated entity is ordinarily soft-deleteable.